config: Allow empty tls.ca_file
Empty ca_file is expected configuration when using a certificate singed
by a trusted CA. This can be organization private CA installed on the
hosts, or a public CA installed on the host by default.

We know that some users want to use such certificates. In older versions
of the proxy, this was hard since we used cert_file as ca_file.  Now
this configuration is easy, but we blocked it.

Tested manually by installing test/conf/ca.pem on a host:

    # cp test/pki/ca.pem /etc/pki/ca-trust/source/anchors/imageio-test-ca.pem
    # update-ca-trust extract
    # rm /etc/pki/ca-trust/source/anchors/imageio-test-ca.pem

Starting a server with empty ca_file:

    $ ./ovirt-imageio -c test/conf
    [server] Starting (pid=548934, version=2.0.5)
    [services] Creating remote.service on port 0
    [services] Securing server (cafile=, certfile=test/pki/cert.pem, keyfile=test/pki/key.pem)
    [services] remote.service listening on port 40631

Now we can access the server without specifying a cafile:

    $ curl -XOPTIONS https://127.0.0.1:40631/images/*
    {"features": ["extents", "zero", "flush"], "unix_socket": "\u000000152"}

Change-Id: I067237bd4b0ea5f10415b7dae44f05e98b2afee3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>