client: Convert to package
Convert the client modules to python package to make packaging and using
easier.

Users need to import only the client:

    from ovirt_imageio import client

    with client.ProgressBar() as pb:
        client.upload(filename, url, progress=pb)

The spec is simplified, including the client python package in
ovirt-imageio-client, and excluding it in ovirt-imageio-common.

Removing the ovirt_imageio.ui module requires update in the
ovirt-engine-sdk examples, and may break users code based on old
examples, but since 4.4 was not released yet, this is our last chance to
change this.

Change-Id: I77ce2be06ecf987558e491fe30e42b689d67560a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>