examples: Improve imageio client example
Rename the example to imageio-client, since this is not a generic http
client.

Improve the usage example and add a ticket example file. Following the
instructions, one can upload and download a raw image using the file
backend.

More work is needed to show how block device can be used, and how to use
the NBD backend.

Change-Id: I334428fb7bd90e031c1907059318fbd54cb92d01
Signed-off-by: Nir Soffer <nsoffer@redhat.com>