setup.py: Update authors and author_email
We had long list of names that do not work on the project any more, with
inactive redhat.com email addresses. Replace the useless list with
"oVirt Authors" and "devel@ovirt.org" so we don't have to update this
again.

If someone wants to see who contributed to the project they can use:

    git shortlog

Or use github:

    https://github.com/oVirt/ovirt-imageio/graphs/contributors

Change-Id: If941cb733d4a0dd5552310164d50dbd9de79e876
Signed-off-by: Nir Soffer <nsoffer@redhat.com>