config: allow keyword__ prefix in configloader
We are going to add default for logger configuration into config module.
One of the attributes of logger configuration is "class" for logger
handler. This is however a keyword in Python and cannot be used in
config class.

Handle such cases by adding "keyword__" prefix to an attribute (with
double underscore to avoid accidental clashes). This prefix will be
added to any key from the loaded config file, which is keyword in
Python. E.g. class key in logger config file mentioned above will
be loaded into attribute keyword__class in object in config module.

Change-Id: I6899184097664aacff21514a0ddf6bb9db96f49c
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>