config: add function for converting object to dict
We want to create initial logger configuration from default in config
module added in previous patch. To be able to load user configuration,
we want to do it using configparser. This accepts initial configuration
as a dictionaty. To be able to convert config classes into dict easily,
add helper function into configload, which converts object to dist and
also handles special case of attributes starting with keyword__ prefix.

Change-Id: I483427853cc8321e2aa3f40b3d73fe78d65e25b5
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>