tests: Fix racy test
We waited until server socket was bound before verifying that the server
dropped privileges, but this racy, and fails randomly in jenkins with:

    istatus = process_status(p.pid)
    pwnam = pwd.getpwnam("nobody")
    assert status["uids"] == {pwnam.pw_uid}
    assert {0} == {65534}

The reason for the failure is that we drop privileges after we bind the
sockets (this way we can bing privileged port).

Fix by waiting until server is listening, which is always after we
finished dropping privileges.

Change-Id: I1db493a810b939c8129d7718f8c1b6360ce4110e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>