server: Separate logger configuration from loading config
Recently we changed the function loading configuration to also configure
the logger. In the next patch we want to add an option to show the
current configuration, and in this case we don't want to configure the
logger. Configuring the logger may fail (e.g. configuration is invalid)
and will prevent reporting of current configuration.

This is also not a good idea in general. For example in server_test.py
we call load_configs() multiple times and this modifies logging
configuration which is global state.

Extract configure_logger() helper and call it after loading the
configuration.

Change-Id: I2c7e3d44812eecde62c9b5e56dc8719b0a864907
Signed-off-by: Nir Soffer <nsoffer@redhat.com>