ops: Remove support for upload without size
Few years ago we added support for upload without specifying size for
vdsm. When we added it we added also support for download without
specifying size, but this was never used and not needed since we always
have backend.size(), and we don't support chunked encoding.

Remove the unneeded support.

Change-Id: I3922c86d8905d705a14795c23011d7e080260376
Signed-off-by: Nir Soffer <nsoffer@redhat.com>