client: Use proxy_url if transfer_url is not accessible
When creating an image transfer, engine reports both transfer_url and
proxy_url. We recommend to use transfer_url for best performance, but
sometimes user need to transfer and image on a machine that can access
ovit-engine, but cannot access the host.

This issue is solved now in ovirt ansible collection by trying to
connect to transfer_url and falling back to proxy_url if transfer_url
is not accessible[1]. This is easier to use and more robust, but it has
to be implemented by all users.

Add a proxy_url argument used as a fallback if transfer_url is not
accessible.

Here is an example usage:

    client.upload(
        filename,
        transfer.transfer_url,
        proxy_url=transfer.proxy_url)

Since this feature was not available before 2.0.7, users should check
client.upload() and client.download() signature before using proxy_url:

    import inspect
    ...
    sig = inspect.signature(client.upload)

    if "proxy_url" in sig.parameters:
        extra = {"proxy_url": transfer.proxy_url}
    else:
        extra = {}

    client.upload(filename, transfer.transfer_url, **extra)

[1] https://github.com/oVirt/ovirt-ansible-collection/pull/35/commits/7c90057b85d937c9ca81eb1f78f59fdba49cb9f4

Change-Id: I5a6b4913d62e34a49ec5ba6677ef578deb0741aa
Bug-Url: https://bugzilla.redhat.com/1839400
Signed-off-by: Nir Soffer <nsoffer@redhat.com>