tests: Test download of raw image
Since the file backend does not support extents, there is no point to
test sparse images. Anything we download becomes preallocated.

When downloading raw image, we can convert it on-the-fly to other qcow2.
The destination should have identical content but because the file
backend does not support extents allocation will be different.

Change-Id: Ia0d658376544ea6a8338f95b543b7acbefdbdea2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>