tests: Test downloading qcow2 image as raw
When downloading snapshots we must use the file backend returning the
actual file data, which is qcow2 bytestream. It may be confusing, but we
must use fmt="raw" in client.download(). We open qemu-nbd using "raw"
format, and it writes the image data as is to the destination file. The
result is an identical qcow2 image.

Since the file backend does not support extents allocation of the
destination file may differ, but in the specific case tested, we get
also identical allocation.

The new test shows that we can use imageio client in
download_disk_snapshot.py example script[1].

[1] https://github.com/oVirt/ovirt-engine-sdk/blob/master/sdk/examples/download_disk_snapshots.py

Change-Id: I39e5dd134f0984fade022d47bce0aa9b7598cc2a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>