backends: http: Add connect_timeout, read_timeout
Add timeout arguments to control client timeouts:

- connect_timeout - the time to wait for socket.connect(). The default
  is 10 seconds since it is unlikely that we want to work with a server
  that takes more than that to connect.

- read_timeout - the time to wait when reading form the socket when
  getting response or response data. The default is 60 seconds since
  server may be block on slow storage, and we don't like to abort the
  request too early.

Previously both settings used the default socket timeout, which may be
too long (several minutes), especially when trying to connect to
inaccessible server.

The default timeouts are used for both client and proxy. We may like to
add the timeouts to daemon configuration later, or maybe to engine so we
can control the timeouts using engine config.

Change-Id: Ibd8dd02a8049800bab7a6a2b97c2a3bbb4ba294f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>