ipv6: support IPv6 in http.Server
Overwrite http.Server.server_bind() method to support IPv6. As the
socket is created in costructor of parent class, we can either
overwrite constructor or overwrite server_bind() method
and recreate the socter. As overwritting TCPServer constructor is
not recommeded according to socketserver.TCPServer documentation,
recreating of the socket was choosen.

Two tests are xfailed, as TCPServer returns tuple of a different size
for IPv4 and IPv6 interfaces. This will be fixed in subsequent patch.

Change-Id: I57357ea4c3cb111efc3dc3f80e38fa64b66b5ef0
Bug-Url: https://bugzilla.redhat.com/1805267
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>