backends: memory: Implement clone()
When using multiple connections, io.copy() needs to create multiple
source and destination backends. Cloning the source and destination
backends allows using the same interface to support single or multiple
connections.

In the past we tried to solve this issue by adding backend factory()
function, but this does not work very well for the http or file
backends. For these backends we like to reuse the information from the
first backend when creating the other backends.

Start by implementing clone() for the memory backend, since it is the
simplest backend.

Change-Id: Ia10eb53b2d0dbdb98b724af9d1677c70aedbf57e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>