backends: http: Separate connection from __init__()
When we clone http backend, we want to skip most of the initialization:
- Read cafile and create SSL context
- Connect using HTTPS
- Resolve server DNS name
- Get OPTIONS
- Initialize server capabilities based on server options
- Switch connection to unix socket if possible.

To make this possible, extract these to _connect() method, and add
connect=True argument so we can create an instance without connecting
it.

When cloning we will copy all the state from the original backend, and
clone the underlying connection.

Change-Id: I053fc5213bb8dbef0cb761b97240760b4bebe33f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>