backends: http: Extract _create_unix_connection()
When we clone a backend connected via unix socket, we want to skip the
handshake using HTTPS and connect directly to the unix socket.

Extract _create_unix_connection(unix_socket) from to be used during
cloning. Rename _create_connection() to _create_tcp_connection() for
consistency.

Both helpers log debug message with the connections details to make it
easier to debug issues. Since we have a debug message for the connection
attempt, we don't need the "Using unix socket" debug log.

Change-Id: I2280c68db9eb6b7e8af2396e0f61ea89db02c221
Signed-off-by: Nir Soffer <nsoffer@redhat.com>