daemon: prefer IPv4 address for control service
When address of control service is resolved, Python resolves the address
as the first working address obtained from getaddrinfo(). This is how
most of the tools work. However, Java has java.net.preferIPv6Addresses
option, which either prefers IPv6 or IPv4 addresses and if the first
one resolved address doesn't work, it fails (see [1] for more details).

Currently, oVirt engine by default prefers IPv4 addresses. Once we
added support for IPv6 into imageio, address of control service can be
now resolved into IPv6 address. Therefore, if daemon resolves control
service address to IPv6 address and listens on IPv6 address, then oVirt
engine fails to connect to control service.

To fix this issue and preserve backward compatibility with engine, add
an option to prefer IPv4 address for control service with the default
set to True. This should work on dual stack when IPv4 will be preferred
as well as in IPv6 environment when there is no IPv4 address and we will
fall back to standard resolution of the address.

[1] https://bugs.openjdk.java.net/browse/JDK-8170568

Change-Id: Ibb624d00723eda2d48a74442ce70e9ecf92b431d
Bug-Url: https://bugzilla.redhat.com/1851707
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>