client: Add write() and read() methods
When you cannot read into client the buffer, ImageClient.read_from() is
not helpful, and using write() is more efficient and easier to use.

For consistency, add a read() method. This can be very inefficient but
may be useful in some cases.

Both methods do not support accessing offset after the end of the image.
This forces clients to pay attention to the offset, but make it much
easier to use the API, since we can never have short writes and reads.

Change-Id: I790890363eada638a4ffc6db1063dc26728d1934
Bug-Url: https://bugzilla.redhat.com/1855047
Signed-off-by: Nir Soffer <nsoffer@redhat.com>