examples: Streaming volumes from libvirt
Libvirt stream is not efficient with imageio, since it does not have a
way get the size of data extents, and it returns small chunks of data.

Add iter_frames() helper, buffering data from libvirt, and yielding data
and hole frames. With this helper streaming data from libvirt is easy
and we get good performance.

Uploading images from libvirt is 2 times slower compared with qemu-nbd.

Change-Id: I063cbf4401c9a156567e293b3d4e452f6aaf23b5
Bug-Url: https://bugzilla.redhat.com/1855047
Signed-off-by: Nir Soffer <nsoffer@redhat.com>