ops: Support cancelling operations
Operations have now a cancel() method. Calling cancel() mark the
operation as canceled. When the operation completes the current I/O it
will check the flag and abort. In normal condition this should take less
than a second.

If operation is blocked on inaccessible storage, it will abort only when
the I/O completes. With NFS this can take several minutes but we don't
have any way to interrupt this.

Change-Id: I34cb5be014da45459f62e6e5e10007c1aca83a66
Bug-Url: https://bugzilla.redhat.com/1524184
Signed-off-by: Nir Soffer <nsoffer@redhat.com>