auth: Support canceling tickets
Canceling a ticket marks the ticket as canceled. All future operations
with this ticket will fail with AuthorizationError.

Canceling a ticket is not enough to remove it, since a ticket may be
used by some connections, and have open file descriptors that may
prevent logical volume deactivation.

After a ticket is marked as canceled, we cancel all ongoing operations
and wait on an event until all connections are removed from the ticket.

If the ticket is still used after a timeout (60 seconds), the cancel
operation will raise errors.TicketCancelTimeout.

To track connections using the ticket, connection context should be
added to the ticket when a connection starts to use a ticket, and
removed from the ticket when a connection is closed.

Ticket.info() reports now the canceled status and the number of
connections using the tickets. This can be used by management
application to wait until a ticket is unused if removing a ticket timed
out.

Change-Id: I81f6e12ea490070593e2b08a36fb94454b3872ea
Bug-Url: https://bugzilla.redhat.com/1524184
Signed-off-by: Nir Soffer <nsoffer@redhat.com>