http: Add connection id
We want to keep connection context in the ticket to allow ticket
cancellation. Add connection id that will be used to lookup connection
context in the ticket.

Update the connection log to show the connection id.

Move the CLOSE log after the context is closed to make it easier to
follow the connection. Now this log is the last log for this connection.

Change-Id: I01d6d87dc75fe03e6f2aac4136c3dfb4484943ea
Bug-Url: https://bugzilla.redhat.com/1524184
Signed-off-by: Nir Soffer <nsoffer@redhat.com>