backends: Keep connection context in the ticket
When creating a backend, we add the context to the ticket using the
connection id.  Because the ticket may be canceled at this point, this
method may fail with AuthrizationError, so we must handle this failure
and close the context.

We register a closer object with the connection, that will remove the
context when the connection is closed. This ensure that the ticket can
be removed when all connections using it are closed.

In images and extents requests handlers, we handle AuthorizationError
when authorizing a ticket, getting backend, and running a ticket. In all
cases we fail with "403 Forbidden", and close the connection. This will
close the connection context and remove the connection from the ticket.

To simplify error handling, we close the connection now on all auth
errors, not only when ticket was canceled. I don't think keeping open
connection for a client trying to do something it is not allowed to do
is good use of resources.

Change-Id: Ia3b1f81ac3b39a6d10cd43522629646fb7e1cf28
Bug-Url: https://bugzilla.redhat.com/1524184
Signed-off-by: Nir Soffer <nsoffer@redhat.com>