examples: sparse-stream: Add meta frame
The example demonstrates:
- Getting all extents from the client using list().
- Using new "data" attribute to calculate data size.
- Using new "data" attribute to simplify handling of non-zero and dirty
  extents when downloading.
- How to extract image virtual size from the last extent.
- How to get dirty extents during incremental backup.
- How to handle zero extents in full and incremental backup.

Add metadata frame with stream metadata like virtual size, data size and
date of the transfer.

When downloading this metadata is useful for allocating space on the
destination storage, reporting progress, or keep data needed to managed
the backup.

During upload this can be useful for validating the destination size,
reporting progress or allocating storage for the image when streaming to
block storage.

Here is example metadata frame from fedora-31 image:

meta 0000000000000000 0000000000000083
{
    "virtual-size": 6442450944,
    "data-size": 1280180224,
    "date": "2020-07-09T23:07:29.924239",
    "incremental": false
}
data 0000000000000000 0000000000010000
<EB>c<90>^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^@^...

Change-Id: I8eeaa94ee5315fe6334ebacfd936a2b7b8581968
Bug-Url: https://bugzilla.redhat.com/1855047
Signed-off-by: Nir Soffer <nsoffer@redhat.com>