io: Simplify copy loop using new "data" attribute
The new "data" attribute can be used to simplify also io.copy() loop.
Previously we had 2 mostly identical copy loops, one for copying data
extents, and for copying dirty extents.

Both loops do the same thing for zero extents with zero=False and dirty
extents with dirty=True, so we can unify this using data=True.

When using dirty extents we must skip non-data extents. When using zero
extents the caller control zeroing. For example when downloading image
to new file, we can always skip non-data extents. But when we upload
file to the server, we must zero since we don't know what is the current
content of the disk.

Change-Id: I6d7d6e9f94225ea8a737a8d8e44968178d59a936
Signed-off-by: Nir Soffer <nsoffer@redhat.com>