checksum: Integrate with ticket.run()
Wrap the checksum computation with an operation, and run the operation
using ticket.run().

This has important properties:

- When the operation is running, the ticket is reported as active, so
  engine will not try to cancel the ticket.

- Cancelling the ticket aborts the operation quickly without waiting
  for connection timeout.

- When the operation completes, the ticket is extended and idle_time is
  reset. This helps engine to track ticket activity.

- The ticket reports operation progress. This may useful only for big
  disk, since engine polls ticket progress every 10 seconds.

When computing checksum of preallocated or empty image we have one
extent and no progress. To get progress we need to split big extents
like we do in io.copy(). I think this should be solved elsewhere, at the
backend level, or as a global helper.

Supporting cancellation and ticket status reports requires too much
code. We need to think how this can simplified.

Change-Id: I809e5769c530118ef1d9ea33ace4b23829ff0e4a
Bug-Url: https://bugzilla.redhat.com/1787906
Signed-off-by: Nir Soffer <nsoffer@redhat.com>