tests: Add helpers for dealing with flaky tests
Some tests are flaky in certain CI environment and should be marked as
xfail. Currently the only issues we have are in oVirt CI.

Add ci module for detecting the CI environment. Currently this just
hides the check for "OVIRT_CI" or "TRAVIS_CI" in os.environ, but is
makes the code little nicer.

Add flaky_in_ovirt_ci marker to make it easy to mark the flaky tests.
This marked use strict=False since the tests fail randomly.

Change-Id: I1b87f2fb2435a71aacb74523485412132b47a560
Signed-off-by: Nir Soffer <nsoffer@redhat.com>