server: Show hostname in starting log
Recently there are too many case when we get daemon log from testers or
users, and it is not clear if the log came from a host or from engine.
The worst case was daemon log form a host that was both engine and host
(unsupported all-in-one installation). Log the hostname in the
"Starting" log to avoid this confusion.

Here is an example log:

2020-08-18 21:58:55,770 INFO    (MainThread) [server] Starting
(hostname=my.server pid=47247, version=2.1.0)

Change-Id: Ib52052d8c29f262095eb974e4deb94eb62ff84e9
Related-To: https://bugzilla.redhat.com/1869034
Signed-off-by: Nir Soffer <nsoffer@redhat.com>