auth: Log when waiting until ticket is unused
When removing a used ticket, we wait until the ticket become unused
before completing the request. Looking at the logs, we can see the
REMOVE log, and the CLOSE log after a while, but it is not clear that
the remove thread is waiting. Promote the debug log to info level to
make this more clear.

Here is an example session:

2020-08-18 22:10:37,284 INFO    (Thread-14) [tickets] [local] REMOVE ticket=eafbe4d0-f651-4fac-9c1d-c85c3ea014c0
2020-08-18 22:10:37,284 INFO    (Thread-14) [auth] Waiting until ticket eafbe4d0-f651-4fac-9c1d-c85c3ea014c0 is unused
...
2020-08-18 22:11:01,707 INFO    (Thread-14) [auth] Ticket eafbe4d0-f651-4fac-9c1d-c85c3ea014c0 was canceled
2020-08-18 22:11:01,707 INFO    (Thread-14) [http] CLOSE connection=14 client=local [connection 1 ops, 24.422921 s] [dispatch 1 ops, 24.422622 s]

Change-Id: I33745e021c8d1b3fdc94dc64e704266528a4730c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>