client: Support download with a backing file
When downloading single snapshot, we need to set the backing file in the
downloaded image. Add backing_file and backing_format options to
client.download().

With this change we can backup all disk snapshots, reconstructing the
image chain. The local chain can replicate the remote backing file
names, or use different names.

This also allows linking incremental backup to the previous backup,
creating an image chain that can be used later to restore the image to
any of the images in the chain.

More work is needed to fix handling of zero extents. Currently they are
skipped instead of writing zeroes. This will expose data from the
backing chain.

Change-Id: I75619412c4be6dff2a4045385d52b78f372d4569
Bug-Url: https://bugzilla.redhat.com/1847090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>