checksum: Always return a dict
Computing a checksum is meaningless without the algorithm and block
size. Change checksum.Operation to return a dict with all the details so
we get dict both from server and client.

This also make the tests easier since we can compare the checksum
retuned from blkhash.checksum(), client.checksum() and server without
reconstructing the dict.

Example output with this change with sdk example script:

$ ./checksum_image.py snap3-full.qcow2
{
  "algorithm": "blake2b",
  "block_size": 4194304,
  "checksum": "0814cf35f454b705af354f1a9e123d9b40a9b1291afc3f5d9c34d7a45d86c87f"
}

Change-Id: I502e479af5d2d2be9396991b9b360aece96fa925
Signed-off-by: Nir Soffer <nsoffer@redhat.com>