io: Fix backend closing in executor.
When using a thread pool we try to reuse the connection backend by
passing the connection backend to the first worker instead of cloning
it. This introduce some issues:

- The backend is closed twice: once by the worker, and again by the
  connection. This is safe since closing backend twice does nothing, but
  dirty. Will break if we try to use the backend after a copy.

- With very small images, the first worker may close the original
  backend before another worker try to clone it. This is unlikely to
  happen with real images but can break the tests.

Fix the issues by introducing a backend wrapper, overriding close() to
disconnect from the wrapped backend instead of closing it. This ensures
that dst is never closed before it is cloned, and that it is closed
exactly once.

Another approach to fix this is to close the initial backend used for
preparing a copy, pass the backends open argument to io.copy(), and open
new backends in the workers without all the hacks. This requires lot of
work, so I'm staring with this minimal change.

Fixes #23

Signed-off-by: Nir Soffer <nsoffer@redhat.com>