ipv6: improve error logging when IPv6 is not supported
Catch an exception when socket is created and log address family and
related stuff.

By default we use "::" special IPv6 address. Creating a socket for such
address would fail on machines where IPv6 is not enabled. To identify
such cases easily, log server address and other related stuff when we
fail to create a socket.

Change-Id: I2608d5e32a6d08e1141220f33ad7195d51805caf
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>