tests: Mark xfail tests with qemu-nbd 6.0.0
The test reveal a real bug, caused by change in qemu-nbd hole reporting.
It reports holes correctly for raw images, but reports them wrongly for
qcow2 images.

If we have this chain:

base: ABC---
top:  -D0---

We except to qemu-nbd to report cluster 3 as zero but not as a hole,
since this cluster is allocated in the top image. This has no effect on
the guest since cluster 3 is read as zero, but when uploading top
imageio will create a hole in the destination image, and this hole
exposes the cluster 3 from base to the guest.

Enhance the tests:

- log the extents in the base and top images to make it easy to debug
  this issue.
- copy both top and base to the destination, instead of coping only top
  and removing the backing file before the compare.
- always compare the entire image both with strict=False to compare guest
  visible content.
- when using qemu-nbd 6.0.0, reporting holes properly, compare also
  allocation.
- Adapt raw file tests to succeed with qemu-nbd 6.0.0. Previously the
  tests assumed the old broken behavior, reporting holes as data.

Mark the affected tests as xfail:
- upload shallow
- download shallow
- extents from qcow2 image
- extents from qcow2 image in ova

Change-Id: I06489f1ba16faa742ebefacacaffc02d2e465613
Signed-off-by: Nir Soffer <nsoffer@redhat.com>