build: Require qemu 6.0.0
Qemu 6.0.0 is available now in Centos Stream advanced virt repo[1]. We
want to consume it now to detect regressions early.

Unlike vdsm, we test only with Centos Stream, so we can require it also
in the CI packages files.

Since qemu-6.0.0 is available on all tested distros, the next step is
removing the conditions checking qemu versions.

[1] http://mirror.centos.org/centos/8-stream/virt/x86_64/advancedvirt-common/

Change-Id: Ibe3e55c6fc9225b342d50616e9604f18e1df3da6
Signed-off-by: Nir Soffer <nsoffer@redhat.com>