tests: Fail on oVirt CI instead of qemu version
The upload hole at start test was mark as failing on qemu < 6.0, but now
that qemu 6.0 is available on the CI, we see that it fail because the CI
environment (likely strange file system).

Change the test to xfail when running in oVirt CI, and log image extents
on failures. The debug logs may explain the failures.

Change-Id: I0b10ff74e4ba29cfaaa58d2614e34f0ff999be75
Signed-off-by: Nir Soffer <nsoffer@redhat.com>