examples: Update sparse-stream example
Now that we report zero extents during incremental backup, we can
optimize incremental backup by downloading only the data extents.

Extent.data changed the semantics - previously it returned True if the
extent was dirty. Now it returns True only if the extent is non-zero, so
it cannot be used to select dirty extents. We filter out the clean
extent before copying instead.

Change-Id: I8191254c1dc116d82539e4252e1d064c17d6b883
Bug-Url: https://bugzilla.redhat.com/1971185
Signed-off-by: Nir Soffer <nsoffer@redhat.com>