tests: Test extents() via nbdutils.extents()
Testing raw nbd.Client.extents() is fragile, since NBD server is allowed
to return short or response. Always test the call wrapped in
nbdutils.extents(), normalizing the output and ensuring that exact range
is returned.

This will also enable considering allocation depth in the response, that
will be implemented in nbdutils.extents().

Change-Id: Ibcc353f367944d6a2a779d82a18a06ffe8f21862
Signed-off-by: Nir Soffer <nsoffer@redhat.com>