nbdutil: Support merging extents
When getting dirty extents, we have both "base:allocation" context and
"qemu:dirty-bitmap:name" context, but we report only the dirty status.
Reporting also zero status would be more useful to callers.

The process depends on having distinct bits in different context, so we
can merge the flags bits from different context.

Add nbdutil.merged() iterator, accepting 2 lists of extents and
generating extents with merged flags.

When merging extent of different length, we consume the shorter extent
save the rest of the longer extent, and repeat with next extent from the
list.

NBD server is allowed to return short reply, so we may have extents
lists describing different byte range. In this case we stop the
iteration once one of the list is consumed, and drop the rest of the
extents from the other list. The caller will have to call
Client.extents() again with the last offset reported.

Change-Id: Ib989a7796248033058facbad3f7a0407e93e6e33
Bug-Url: https://bugzilla.redhat.com/1971185
Bug-Url: https://bugzilla.redhat.com/1971182
Signed-off-by: Nir Soffer <nsoffer@redhat.com>