tests: Rewrite nbdutil extents tests
The previous tests were complicated and rigid, and I cold not find any
way to test allocation depth and dirty bitmaps.

This change introduce a much simpler client that can be configured with
extents lists for base:allocation, qemu:allocation-depth, and
qemu:dirty-bitmap:name.

Unfortunately, to return extent reply from the configured extents, the
client need to be pretty smart. It have to search in the extent list and
return the extents overlapping the requested range. This includes also
clipping extents starting before the requested range, or extents ending
after the requested range.

Additionally, the client can be configured to return short reply. I'm
not sure qemu NBD server returns short replies, but the NBD protocol
allows it, so we must be prepared to handle this.

Since the client code is pretty tricky, we have new tests verifying that
that the client behaves correctly. These tests also demonstrate why
merging extents from different meta contexts is tricky.

All the client and extents tests use the same client, configured to test
several interesting uses cases that were not tested before. This client
is tested with all the combinations tested before - getting all extents,
getting extents at the start, middle and end of the image, and getting
zero and dirty extents.

Change-Id: Ibc7c9a59a3f64b6727075a453b55eee74b81a8e6
Bug-Url: https://bugzilla.redhat.com/1971185
Bug-Url: https://bugzilla.redhat.com/1971182
Signed-off-by: Nir Soffer <nsoffer@redhat.com>