docker: Workaround for fedora 34 image
When building fedora 34 image on quay.io, /lib/modules/*/modules.dep is
missing. This breaks virt-builder, failing the backup tests. I filed a
quay.io bug[1] for this, but I don't think we it will be resolved soon.

It seems that the issue is that depmod is not run as it should when
installing kernel-code package. Reinstalling the package or running
depmod manually fixes the issue. Lets run demmod since it is much
faster.

[1] https://issues.redhat.com/browse/PROJQUAY-2130

Change-Id: I7fc7adae0c57a7d0539178f812a69e63d34c8eb3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>