qemu_nbd: Support for old qemu-nbd on RHEL
RHEL ships an older version of qemu-nbd (4.2.0) that does not support
the --allocation-depth option. Previously we always use this option, so
running the imageio client with qemu-nbd < 5.2.0 would fail, even for
operation that do not require this option.

Add this option only on qemu-nbd >= 5.2.0 so imageio client is
compatible with older qemu-nbd when not using backing_chain=False.

Using backing_chain=False will fail on qemu-nbd < 5.2.0, since it
requires --allocation-depth.

Change-Id: I15f574f132c808c6816a9fbff64be30cf56909fd
Signed-off-by: Nir Soffer <nsoffer@redhat.com>