spec: Fix qemu-img requirement
The requirement was missing the epoch, which makes it ineffective,
accepting older versions.

Change the version to 4.2.0 since we can work now with this version,
since --allocation-depth is used only on qemu-nbd >= 5.2.0.

Because the requirement was broken, this practically does not change
the behavior on RHEL 8.4, but it makes it clear that we expect the
package to work with older qemu-nbd.

Change-Id: I5facdd62d44895fc12e0e003a8d587348656be9e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>