test: Fix wrong test for shallow upload
When uploading raw base image, we cannot detect unallocated area, since
qemu-nbd reports unallocated area in raw image as data instead of a
hole. So we treat both zero and unallocated area as zero area, and the
target image allocation is not the same as the source image.

This failure was hidden since the test wrongly skip the test with
qemu-img 6.0.0. Now that qemu-img 6.1.0 is available, the error is
revealed.

Skip the allocation check when the base image is raw.

Change-Id: I751927b0d6b9479965fd23c38a2af9d0a4a678bf
Signed-off-by: Nir Soffer <nsoffer@redhat.com>