go: Improve code style and documentation
Fix various issues reported by "golint" and "go vet":

- should not use dot imports
- comment on exported const MAX_EXTENT should be of the form "MAX_EXTENT ..."
- don't use ALL_CAPS in Go names; use CamelCase
- exported function Connect should have comment or be unexported
- exported func Connect returns unexported type *nbd.backend, which can be annoying to use
- exported type Extent should have comment or be unexported
- exported func Connect returns unexported type *http.backend, which can be annoying to use
- *ovirt.org/imageio.Extent composite literal uses unkeyed fields

This was not fixed, since there is nothing to comment about the KiB,
MiB, and GiB constants.

- exported const KiB should have comment (or a comment on this block) or be unexported

golint is deprecated[1], but most of its checks are useful. Some changes
like avoiding dot imports make the code more verbose but it is nice that
the linter does not complain.

[1] https://github.com/golang/go/issues/38968

Change-Id: I253cfa16cb4b227d1cdb18ca2300c4d1d077f08b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>