ovirt-img: Prepare for extents merging
To merge NBD extents, need to compare the flags, which are dropped now
when creating Extent from every pair. However the http backend, we
already get merged extents. In both cases we don't really need to get a
slice of extents; we need a way to iterate over them.

Change the Backend interface to return new ExtentsResult interface. This
interface supports Next() and Value(), which can be used with for loop
to iterate over the extents.

The NBD ExtentsResult implementation keeps the (length, flags) pairs
returned from the NBD server. During iteration, each pair is converted
to Extent. This allows merging of extents with same flags or different
meta context in the future.

Change-Id: I79a78dfb0f3d7f90e8d159f4835926c4bd81c626
Signed-off-by: Nir Soffer <nsoffer@redhat.com>