ovirt-img: Fix handling of short or long reply
NBD server may return a short or long reply:

- short reply: one or more extents ending before the requested range.
- long reply: all extents, the last extent ends after the requested
  range.

In the second case, we want to consume all the extents to minimize the
number of calls to the server, and to avoid duplicate work on the server
side. In both cases, we need to continue where the last extent ended.

Change-Id: I8331bee57724f28e567bc1308b2e00a7d6686ad4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>