server: Make load_config usable from other modules
The implementation for loading server configuration belongs to the
server module, but it is useful for other modules that want to connect
with the server.

We already have a public server.load_config() function, but it accepted
an argument parser object, while it need just the configuration
directory (/etc/ovirt-imageio). Change the function to accept the
configuration directory.

Callers of server.load_config() need to know the default configuration
directory, available at server.DEFAULT_CONF_DIR.

Change-Id: I0f0b71eb91fa0a866c32353b4c4718ea98d41078
Signed-off-by: Nir Soffer <nsoffer@redhat.com>