WIP: docs: Administration guide
Start new document describing how to configure the ovirt-imageio service
to expose images. In oVirt the vdsm service manage the ovirt-imageio
service, but it is possible to mange the ovirt-imageio service directly
using the admin library or ovirt-imageioctl command.

This guide is will also be helpful for developers to understand how
ovirt-imageio works.

Preview the guide here:
https://github.com/nirs/ovirt-imageio/blob/admin/docs/admin.md

Change-Id: Ibc8ccfeabe3f57ccab5fc7d9d3f231b112048b4b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>