ovirt-img: Use JSON output
Output JSON instead of inventing a new text format. JSON is useful both
for humans and machines, so it should be the default format.

To have nicer output, minimize memory usage, and speed up formatting, do
not use `encoding/json`, but format the JSON manually.

Example output:

    $ ./ovirt-img map $URL
    [{"start": 0, "length": 65536, "zero": false},
     {"start": 65536, "length": 983040, "zero": true},
     {"start": 1048576, "length": 65536, "zero": false},
     ...
     {"start": 5564727296, "length": 48562176, "zero": false},
     {"start": 5613289472, "length": 829095936, "zero": true},
     {"start": 6442385408, "length": 65536, "zero": false}]

This format is easy to read and play well with standard line based
tools.

Change-Id: I762e6e3b6b0daecc2906ab7f8ff568a7df178a3d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>