nbd: Cache export size
We need the image size in many places and using nbd_get_size() makes the
code more complicated, having to handle error that will never happen.
This also adds unneeded overhead calling into libnbd; calling to C
functions is not free.

The image size cannot change while we access it, so it is correct to
cache the size.

Change-Id: Ifed6a743d750ee34c7d515a60dff76dc6694f8fe
Signed-off-by: Nir Soffer <nsoffer@redhat.com>