docker: Remove the dnf update phase
This makes building the container slower and less predicable. It also
does not help to get update containers, since quay.io builds the
container only if the FROM layer was changed, or if the docker file
changed. We will consume the content of the FROM container as is now.

While touching the centos-9 container, change to use centos:stream9
instead of stream9-development[1]

[1] https://ovirt.org/develop/developer-guide/migrating_to_github.html

Change-Id: I01fee8ae5ec1eae731db948097b7754e8f493ce9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>