github: Add rpm job
Add job building rpms and uploading built artifacts with all container
images. For OST we need only centos-* rpms, but we want to keep this
project working also on fedora.

The rpm job creates a zip file with a yum repository for every distro
(e..g rpm-centos-8.zip). To run OST with the built rpms, a developer
will use the link the zip in a OST build parameter:

    https://github.com/nirs/ovirt-imageio/suites/4535392764/artifacts/121520882

Here is an example build for preview:
https://github.com/nirs/ovirt-imageio/actions/runs/1531658722

Change-Id: I60c4f67171df4217b218546dceb1422c36bcd03b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>