auth: Cancel ticket faster
Previously we marked the ticket as canceled immediately, but if the
ticket had registered connections, we waited until the connections are
closed, or the timeout expires.

This mechanism was OK when connection timed out after 60 seconds of
inactivity, but we want to allow much longer inactivity timeout, and
this means that canceling a ticket can be too slow.

I found out that recent virt-v2v try to finalize a transfer without
closing the connection the imageio. This result in timeout finalizing
the transfer, since virt-v2v want sets a very large inactivity timeout
(3600 seconds).

We also had know about 2 other buggy clients (ansible upload, qe image
transfer tests) that tried to finalize a transfer without closing the
connection.

Waiting until connections time out was not helpful to anyone, since once
a ticket is marked as cancelled, any request will fail immediately with
authorization error.

Change the cancel flow to wait for ongoing operations instead of
connections. This means that ticket can be canceled immediately if there
are no ongoing operations (e.g. virt-v2v), or we wait very short time
until all ongoing operations finish.

If there are idle connections when ticket is canceled, they will remain
until the connection times out, or the user sends a request. This doe
not affect the server in any way except consuming resources. We can
check later if there is a good way to abort these connections
immediately.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>