auth: Close connections context in cancel()
Once ticket is marked as cancelled, no new operation can start, and new
connections cannot register with the ticket.

If there are no ongoing operations, we need to close connections backend
and buffer before we remove the ticket.

If we have ongoing operations, we need to wait until they finish. The
operations will discover that the ticket was canceled, close their
context and the underlying connection.

However if we had idle connections, we need to close their context when
we finish the wait.

When we close connection context, we don't remove the context from the
ticket, since this will cause the connection to create a new context on
the next request. The context will be removed when the connection try to
start a new operation.

If cancelling the ticket timed out, the user need to poll the ticket
"active" property. When the ticket becomes inactive, the user must
delete the ticket again.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>