backends/ndb: Avoid double close logs
If the backend of an idle connection is closed when ticket is canceled,
the backend is closed again when the idle connection is closed. Closing
backend multiple times is designed to be safe, but we should really log
only once, when actually closing the nbd client.

The same behavior is already implemented in the file backend.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>